+++ NEWS.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - `EasyFITS` replaced by `AstroFITS`.
+- `MultivariateOnlineStatistics` replaced by `OnlineSampleStatistics`.
 
 ## Version v0.3.5